requirements.txt
google-genai>=1.7.0
sqlite-vec>=0.0.5
tiktoken>=0.5.0
tqdm>=4.66.0
numpy>=1.24.0
python-dotenv>=1.0.0